+++ src/is-json.ts
@@ -1,9 +1,9 @@
-import hasJsonProtoype from "./utils/has-json-proto.js";
+import { hasJsonProtoype, hasArrayPrototype } from "./utils/has-json-proto.js";
 
 /**
  * Checks if a value can be considered a plain JSON object
  * @param value The value to be evaluated
  */
 export default function isJSON(value: any) {
-  return value !== undefined && value !== null && hasJsonProtoype(value);
+  return value && (hasJsonProtoype(value) || hasArrayPrototype(value));
 }