+++ src/utils/has-json-proto.ts
@@ -1,6 +1,10 @@
 const JSON_PROTOTYPE = Object.getPrototypeOf({});
+const ARRAY_PROTOTYPE = Object.getPrototypeOf([]);
 
-const hasJsonProtoype = (value: any) =>
+/** Checks if some value contains a JSON `object` prototype */
+export const hasJsonProtoype = (value: any) =>
   Object.getPrototypeOf(value) === JSON_PROTOTYPE;
 
-export default hasJsonProtoype;
+/** Checks if some value contains a JSON `array` prototype */
+export const hasArrayPrototype = (value: any) =>
+  Object.getPrototypeOf(value) === ARRAY_PROTOTYPE;